remove App.test.js
This was actually breaking due to the test runner 'finishing' before the data from Apollo was resolved. I don't see a way of fixing this in unit testing, since there's no way to "wait" for a component to be fully fetched and the App.js component in its current form is meant to be completely opaque i.e just plug and play.

It's likely that the correct answer would be to actually only use App.js in the browser and therefore never test it in unit tests (only in acceptance) and therefore using SSR tech for testing